Add files via upload
The Weather Forecast Web Page is a simple yet handy tool for checking the current weather conditions. It displays real-time data such as temperature, humidity, and wind speed, allowing users to stay updated on the current weather situation. Built with HTML, CSS, and JavaScript, the webpage offers a clean and intuitive interface for quick access to essential weather information.